+++ packages/dataviews/src/types.ts
@@ -44,10 +44,17 @@ export type Operator =
 
 export type ItemRecord = Record< string, unknown >;
 
+export type FieldType = 'text';
+
 /**
  * A dataview field for a specific property of a data type.
  */
 export type Field< Item > = {
+	/**
+	 * Type of the fields.
+	 */
+	type?: FieldType;
+
 	/**
 	 * The unique identifier of the field.
 	 */
@@ -57,6 +64,11 @@ export type Field< Item > = {
 	 * The label of the field. Defaults to the id.
 	 */
 	header?: string;
+
+	/**
+	 * Placeholder for the field.
+	 */
+	placeholder?: string;
 
 	/**
 	 * Callback used to render the field. Defaults to `field.getValue`.
@@ -130,6 +142,13 @@ export type NormalizedField< Item > = Field< Item > & {
 export type Fields< Item > = Field< Item >[];
 
 export type Data< Item > = Item[];
+
+/**
+ * The form configuration.
+ */
+export type Form = {
+	visibleFields?: string[];
+};
 
 /**
  * The filters applied to the dataset.
@@ -232,7 +251,7 @@ interface ViewBase {
 	/**
 	 * The hidden fields.
 	 */
-	hiddenFields?: string[];
+	fields?: string[];
 }
 
 export interface ViewTable extends ViewBase {
@@ -387,7 +406,7 @@ export interface ViewBaseProps< Item > {
 	fields: NormalizedField< Item >[];
 	getItemId: ( item: Item ) => string;
 	isLoading?: boolean;
-	onChangeView( view: View ): void;
+	onChangeView: ( view: View ) => void;
 	onSelectionChange: SetSelection;
 	selection: string[];
 	setOpenedFilter: ( fieldId: string ) => void;