+++ packages/edit-site/src/components/page-patterns/search-items.js
@@ -24,11 +24,42 @@ import {
 } from '../../utils/constants';
 
 // Default search helpers.
-const defaultGetName = ( item ) =>
-	item.type !== TEMPLATE_PART_POST_TYPE ? item.name || '' : '';
-export const defaultGetTitle = ( item ) =>
-	typeof item.title === 'string' ? item.title : item.title.rendered;
-const defaultGetDescription = ( item ) => item.description || '';
+const defaultGetName = ( item ) => {
+	if ( item.type === PATTERN_TYPES.user ) {
+		return item.slug;
+	}
+
+	if ( item.type === TEMPLATE_PART_POST_TYPE ) {
+		return '';
+	}
+
+	return item.name || '';
+};
+
+export const defaultGetTitle = ( item ) => {
+	if ( typeof item.title === 'string' ) {
+		return item.title;
+	}
+
+	if ( item.title && item.title.rendered ) {
+		return item.title.rendered;
+	}
+
+	if ( item.title && item.title.raw ) {
+		return item.title.raw;
+	}
+
+	return '';
+};
+
+const defaultGetDescription = ( item ) => {
+	if ( item.type === PATTERN_TYPES.user ) {
+		return item.excerpt.raw;
+	}
+
+	return item.description || '';
+};
+
 const defaultGetKeywords = ( item ) => item.keywords || [];
 const defaultHasCategory = () => false;