+++ packages/token-list/src/index.ts
@@ -193,6 +193,7 @@ export default class TokenList {
 		return true;
 	}
 
+	/* eslint-disable @typescript-eslint/no-unused-vars */
 	/**
 	 * Returns true if `token` is in the associated attribute’s supported
 	 * tokens. Returns false otherwise.
@@ -207,4 +208,5 @@ export default class TokenList {
 	supports( _token: string ): boolean {
 		return true;
 	}
+	/* eslint-enable @typescript-eslint/no-unused-vars */
 }