+++ packages/dataviews/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+
+### Breaking Changes
+
+- Replace the `hiddenFields` property in the view prop of `DataViews` with a `fields` property that accepts an array of visible fields instead.
+
+### New features
+
+-   Added a new `DataForm` component to render controls from a given configuration (fields, form), and data.
+
+### Internal
+
+-   Method style type signatures have been changed to function style ([#62718](https://github.com/WordPress/gutenberg/pull/62718)).
+
 ## 2.2.0 (2024-06-26)
 
 ## 2.1.0 (2024-06-15)